[deprecation] Move feign encoders and decoders out of the "feign" package
Our style guide states that our code must not live in other packages:
> The package name of internal as well as open source projects
> starts with `com.palantir.<project name>`
https://github.com/palantir/gradle-baseline/blob/develop/docs/java-style-guide/readme.md#package-statement

Deprecation: Classes which were previously in the `feign` package
have been migrated into `com.palantir.conjure.java.client.jaxrs.feignimpl`.
These are implementation details of the conjure-java-jaxrs-client library,
however they are still accessable to avoid blocking existing consumers.
We have provided deprecated classes in the `feign` package which shim to
the new location for now.

Classes originally existed in the feign package to use the
package-private `Types.getRawType` utility. This change updates
those uses to a new internal utility `RawTypes.get` which is
implemented using guava TypeToken.getRawType.
