+++ conjure-java-jaxrs-client/src/main/java/com/palantir/conjure/java/client/jaxrs/feignimpl/RawTypes.java
@@ -1,0 +1,32 @@
+/*
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.palantir.conjure.java.client.jaxrs.feignimpl;
+
+import com.google.common.reflect.TypeToken;
+import java.lang.reflect.Type;
+
+/**
+ * Utility functionality to detect raw types.
+ */
+final class RawTypes {
+
+    /** Attempts to extract the raw type class from a {@link Type}. */
+    static Class<?> get(final Type type) {
+        return TypeToken.of(type).getRawType();
+    }
+
+    private RawTypes() { }
+}
