+++ conjure-java-jaxrs-client/src/test/java/com/palantir/conjure/java/client/jaxrs/feignimpl/TextDelegateEncoderTest.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright 2017 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package feign;
+package com.palantir.conjure.java.client.jaxrs.feignimpl;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -22,6 +22,7 @@ import static org.mockito.Mockito.verifyZeroInteractions;
 
 import com.google.common.collect.Maps;
 import com.google.common.net.HttpHeaders;
+import feign.RequestTemplate;
 import feign.codec.Encoder;
 import java.util.Arrays;
 import java.util.Collection;
@@ -30,7 +31,7 @@ import javax.ws.rs.core.MediaType;
 import org.junit.Before;
 import org.junit.Test;
 
-public final class ConjureTextDelegateEncoderTest {
+public final class TextDelegateEncoderTest {
 
     private RequestTemplate requestTemplate;
     private Map<String, Collection<String>> headers;
@@ -41,7 +42,7 @@ public final class ConjureTextDelegateEncoderTest {
     public void before() {
         delegate = mock(Encoder.class);
         headers = Maps.newHashMap();
-        textDelegateEncoder = new ConjureTextDelegateEncoder(delegate);
+        textDelegateEncoder = new TextDelegateEncoder(delegate);
         requestTemplate = new RequestTemplate();
     }
